fix: 기존 int id를 Long id로 변경

+++ src/main/java/org/jenga/dantong/post/model/entity/Post.java
@@ -1,28 +1,14 @@
 package org.jenga.dantong.post.model.entity;
 
-import jakarta.persistence.CascadeType;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToMany;
-import jakarta.persistence.Table;
-import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.List;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
+import jakarta.persistence.*;
+import lombok.*;
 import org.jenga.dantong.global.base.BaseEntity;
 import org.jenga.dantong.survey.model.entity.Survey;
 import org.jenga.dantong.user.model.entity.User;
+
+import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
 
 @Table(name = "post")
 @NoArgsConstructor
@@ -36,7 +22,7 @@ public class Post extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "post_id", insertable = false, updatable = false)
-    private int postId;
+    private Long postId;
 
     @ManyToOne
     @JoinColumn(name = "user_id")
